:construction: Tag IDs int -> long

+++ Toggl.Foundation/Models/FoundationModels.cs
@@ -127,7 +127,7 @@ namespace Toggl.Foundation.Models
 
         public IList<string> Tags { get; }
 
-        public IList<int> TagIds { get; }
+        public IList<long> TagIds { get; }
 
         public DateTimeOffset At { get; }
